+++ contracts/interfaces/IDXswapPair.sol
@@ -41,7 +41,7 @@ interface IDXswapPair {
     function price0CumulativeLast() external view returns (uint);
     function price1CumulativeLast() external view returns (uint);
     function kLast() external view returns (uint);
-    function swapFee() external view returns (uint8);
+    function swapFee() external view returns (uint32);
 
     function mint(address to) external returns (uint liquidity);
     function burn(address to) external returns (uint amount0, uint amount1);
@@ -50,5 +50,5 @@ interface IDXswapPair {
     function sync() external;
 
     function initialize(address, address) external;
-    function setSwapFee(uint8) external;
+    function setSwapFee(uint32) external;
 }
